Add mobile build artifacts for iOS and Android (#127)
* Add mobile build artifacts for iOS and Android

- Add GitHub Actions workflow to build XCFramework (iOS) and AAR (Android)
  and publish them to GitHub releases on tag pushes
- Add scripts/build-ios.sh and scripts/build-android.sh for local builds
- Add android/ project (Gradle library module with Kotlin JNI wrapper)
- Add include/shoes.h C header for FFI consumers
- Gitignore /output build directory


* Fix Android JNI build: pin jni to 0.21, use --locked

The committed Cargo.toml had `jni = "*"` which resolved to jni 0.22.4 in
CI. jni 0.22 renamed JNIEnv to EnvUnowned, breaking the Android FFI code
written for the 0.21 API.

- Pin jni to "=0.21" (resolves to 0.21.1 via Cargo.lock)
- Add --locked to cargo ndk to enforce the lock file in CI


* Migrate Android JNI bindings to jni 0.22

jni 0.22 split JNIEnv into two types:
- EnvUnowned<'local> — FFI-safe type for native method signatures
- Env<'local> — high-level wrapper with get_string, new_string, etc.

Changes:
- Replace JNIEnv with EnvUnowned in all function signatures
- Use EnvUnowned::with_env() to access the Env inside each function
- Use Outcome enum (Ok/Err/Panic) from into_outcome() for error handling
- Replace new_global_ref return type: GlobalRef -> Global<JObject<'static>>
- Replace attach_current_thread() guard pattern with callback closure
- Replace &str method names/sigs with jni_str!/jni_sig! compile-time macros


* Auto-generate iOS C header with cbindgen

Replace hand-written include/shoes.h with cbindgen-generated output.
The header is now derived from doc comments on extern "C" functions
in src/ffi/ios.rs, keeping it in sync with the Rust source.

- Add cbindgen.toml (exports only ProtectSocketCallback + FFI functions)
- build-ios.sh runs cbindgen before building the XCFramework
- CI installs cbindgen in the iOS job


* Fix clippy and rustfmt CI failures

- Reorder jni imports in ffi/android.rs to match rustfmt order
- Remove unnecessary `as usize` cast in shadow_tls_server_handler.rs
- Collapse match guard in config/validate.rs (clippy::collapsible_match)


---------

+++ src/shadow_tls/shadow_tls_server_handler.rs
@@ -979,8 +979,7 @@ async fn setup_local_handshake(
     write_all(&mut server_stream, server_hello_frame).await?;
 
     // The server sends multiple frames after ServerHello; process remaining data.
-    let remaining_server_data_len =
-        server_data_len - TLS_HEADER_LEN - server_hello_payload_size as usize;
+    let remaining_server_data_len = server_data_len - TLS_HEADER_LEN - server_hello_payload_size;
     if remaining_server_data_len > 0 {
         server_data.copy_within(
             TLS_HEADER_LEN + server_hello_payload_size
